event sourcing for snapshots, start deprecating activeScene
